improvement: HomeFeed persistence and TweetActions
This commit introduces a number of improvements to the HomeFeed and TweetActions components:

~ Refactor HomeFeed: To maintain the user's selected feed (ForYouFeed or FollowingFeed) even after a page reload, we added logic to save the 'isForYou' state to localStorage. This ensures that the user experience remains consistent even across page reloads.

~ Refactor TweetActions: The component was normalized to make it more readable and maintainable. This involved breaking down complex structures into simpler, more manageable parts. 

~ Add New Components: New components were introduced as part of the normalization process. These components help to simplify the overall structure of the app and make the codebase more modular and easier to understand.
